# server/openapi.json
{
	"components": {
		"schemas": {
...
			"CleanRequest": {
				"description": "CleanRequest schema",
				"properties": {
					"mode": {
						"$ref": "#/components/schemas/ModeDTO"
					},
					"rooms": {
						"items": {
							"type": "string"
						},
						"type": [
							"array",
							"null"
						]
					}
				},
				"required": [
					"mode",
					"rooms"
				],
				"type": "object"
			},
...
			"ErrorItem": {
				"properties": {
					"more": {
						"additionalProperties": {
							"description": "Additional information about the error"
						},
						"description": "Additional information about the error",
						"type": [
							"object",
							"null"
						]
					},
					"name": {
						"description": "For example, name of the parameter that caused the error",
						"type": "string"
					},
					"reason": {
						"description": "Human readable error message",
						"type": "string"
					}
				},
				"required": [
					"name",
					"reason"
				],
				"type": "object"
			},
			"HTTPError": {
				"description": "HTTPError schema",
				"properties": {
					"detail": {
						"description": "Human readable error message",
						"type": "string"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/ErrorItem"
						},
						"type": [
							"array",
							"null"
						]
					},
					"instance": {
						"type": "string"
					},
					"status": {
						"description": "HTTP status code",
						"example": 403,
						"type": "integer"
					},
					"title": {
						"description": "Short title of the error",
						"type": "string"
					},
					"type": {
						"description": "URL of the error type. Can be used to lookup the error in a documentation",
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"MQTTConfigDTO": {
				"description": "MQTTConfigDTO schema",
				"properties": {
					"base_topic": {
						"type": "string"
					},
					"broker": {
						"type": "string"
					},
					"connected": {
						"type": "boolean"
					},
					"discovery_prefix": {
						"type": "string"
					},
					"has_password": {
						"type": "boolean"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"base_topic",
					"broker",
					"connected",
					"discovery_prefix",
					"has_password",
					"username"
				],
				"type": "object"
			},
			"MQTTConfigInput": {
				"description": "MQTTConfigInput schema",
				"properties": {
					"base_topic": {
						"type": "string"
					},
					"broker": {
						"type": "string"
					},
					"discovery_prefix": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"base_topic",
					"broker",
					"discovery_prefix",
					"password",
					"username"
				],
				"type": "object"
			},
...
			"ModeDTO": {
				"properties": {
					"fan_level": {
						"type": "integer"
					},
					"times": {
						"type": "integer"
					},
					"type": {
						"type": "string"
					},
					"water_level": {
						"type": "integer"
					}
				},
				"required": [
					"fan_level",
					"times",
					"type",
					"water_level"
				],
				"type": "object"
			},
			"OK": {
				"description": "OK schema",
				"properties": {
					"ok": {
						"type": "boolean"
					}
				},
				"required": [
					"ok"
				],
				"type": "object"
			},
			"Option": {
				"properties": {
					"label": {
						"type": "string"
					},
					"value": {
						"type": "integer"
					}
				},
				"required": [
					"label",
					"value"
				],
				"type": "object"
			},
...
			"RoomDTO": {
				"description": "RoomDTO schema",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object"
			},
...
			"SelfCleanRequest": {
				"description": "SelfCleanRequest schema",
				"properties": {
					"action": {
						"type": "integer"
					}
				},
				"required": [
					"action"
				],
				"type": "object"
			},
			"Setting": {
				"properties": {
					"default": {
						"type": "integer"
					},
					"icon": {
						"type": "string"
					},
					"key": {
						"type": "string"
					},
					"kind": {
						"type": "string"
					},
					"max": {
						"type": "integer"
					},
					"min": {
						"type": "integer"
					},
					"name": {
						"type": "string"
					},
					"options": {
						"items": {
							"$ref": "#/components/schemas/Option"
						},
						"type": [
							"array",
							"null"
						]
					},
					"prop": {
						"type": "integer"
					}
				},
				"required": [
					"default",
					"icon",
					"key",
					"kind",
					"name",
					"prop"
				],
				"type": "object"
			},
			"SettingsDTO": {
				"description": "SettingsDTO schema",
				"properties": {
					"schema": {
						"items": {
							"$ref": "#/components/schemas/Setting"
						},
						"type": [
							"array",
							"null"
						]
					},
					"values": {
						"additionalProperties": {
							"type": "integer"
						},
						"type": [
							"object",
							"null"
						]
					}
				},
				"required": [
					"schema",
					"values"
				],
				"type": "object"
			},
			"SettingsInput": {
				"description": "SettingsInput schema",
				"properties": {
					"values": {
						"additionalProperties": {
							"type": "integer"
						},
						"type": [
							"object",
							"null"
						]
					}
				},
				"required": [
					"values"
				],
				"type": "object"
			},
...
			"StateDTO": {
				"description": "StateDTO schema",
				"properties": {
					"battery_level": {
						"type": "integer"
					},
					"charging": {
						"type": "boolean"
					},
					"cloud_connected": {
						"type": "boolean"
					},
					"device_name": {
						"type": "string"
					},
					"docked": {
						"type": "boolean"
					},
					"error_code": {
						"type": "integer"
					},
					"fan_speed": {
						"type": "string"
					},
					"run_state": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"working_status": {
						"type": "integer"
					}
				},
				"required": [
					"battery_level",
					"charging",
					"cloud_connected",
					"device_name",
					"docked",
					"error_code",
					"fan_speed",
					"run_state",
					"state",
					"working_status"
				],
				"type": "object"
			},
...
		}
	},
	"info": {
...
		"title": "OpenAPI",
		"version": "0.0.1"
	},
	"openapi": "3.1.0",
	"paths": {
...
		"/api/clean": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.func3`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "startClean",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/CleanRequest"
							}
						}
					},
					"description": "Request body for webapi.CleanRequest",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func3",
				"tags": [
					"api"
				]
			}
		},
		"/api/config/mqtt": {
			"get": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.func12`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "getMqttConfig",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MQTTConfigDTO"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/MQTTConfigDTO"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func12",
				"tags": [
					"api"
				]
			},
			"put": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.func13`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "setMqttConfig",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/MQTTConfigInput"
							}
						}
					},
					"description": "Request body for webapi.MQTTConfigInput",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MQTTConfigDTO"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/MQTTConfigDTO"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func13",
				"tags": [
					"api"
				]
			}
		},
		"/api/dock": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.cmd.func20`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "dock",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func20",
				"tags": [
					"api"
				]
			}
		},
		"/api/locate": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.cmd.func28`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "locate",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func28",
				"tags": [
					"api"
				]
			}
		},
...
		"/api/pause": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.cmd.func22`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "pause",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func22",
				"tags": [
					"api"
				]
			}
		},
		"/api/resume": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.cmd.func24`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "resume",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func24",
				"tags": [
					"api"
				]
			}
		},
		"/api/rooms": {
			"get": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.func2`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "getRooms",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RoomDTO"
									},
									"type": "array"
								}
							},
							"application/xml": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RoomDTO"
									},
									"type": "array"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func2",
				"tags": [
					"api"
				]
			}
		},
		"/api/self-clean": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.func9`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "selfClean",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/SelfCleanRequest"
							}
						}
					},
					"description": "Request body for webapi.SelfCleanRequest",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func9",
				"tags": [
					"api"
				]
			}
		},
		"/api/settings": {
			"get": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.func10`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "getSettings",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SettingsDTO"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/SettingsDTO"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func10",
				"tags": [
					"api"
				]
			},
			"put": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.func11`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "setSettings",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/SettingsInput"
							}
						}
					},
					"description": "Request body for webapi.SettingsInput",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func11",
				"tags": [
					"api"
				]
			}
		},
		"/api/state": {
			"get": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.func1`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "getState",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StateDTO"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/StateDTO"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func1",
				"tags": [
					"api"
				]
			}
		},
		"/api/stop": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.NewServer.cmd.func26`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "stop",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func26",
				"tags": [
					"api"
				]
			}
		},
...
	},
...
}